backend/requirements.txt
aioredis
asyncpg
boto3==1.28.49
deepface==0.0.79
fastapi==0.100.0
fastapi_plugins==0.12.0
fer==22.5.1
huggingface_hub==0.23.5
moviepy==1.0.3
numpy==1.24.2
openai==0.28.0
passlib==1.7.4
Pillow==10.0.1
pinecone==5.0.1
pydantic==1.10.11
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart
psycopg2-binary
pytz==2023.3.post1
Requests==2.31.0
scikit_learn==1.3.0
scipy==1.10.1
SQLAlchemy==1.4.48
transformers==4.27.4
tensorflow==2.11.1
sentencepiece
sentence-transformers==2.2.2
